fix coding style issues
