+++ data/hiChat.txt
@@ -1,5 +1,3 @@
-[D] [ ] return book           (by: Dec 2 2019 18:00)
+[D] [ ] return book            (by: Dec 2 2019 18:00)
 [T] [ ] play 
-[T] [ ] haha 
-[E] [ ] project meeting  (from:   Mon 2pm  to: 4pm )
-[T] [ ] heihei 
+[E] [ ] project meeting  (from:    Mon 2pm  to: 4pm )
