+++ docs/README.md
@@ -1,4 +1,4 @@
-# hiChat.HiChat User Guide
+# hichat.HiChat User Guide
 
 // Update the title above to match the actual product name
