+++ src/main/java/hichat/DialogBox.java
@@ -1,4 +1,4 @@
-package hiChat;
+package hichat;
 
 import javafx.geometry.Pos;
 import javafx.scene.control.Label;
