+++ src/main/java/hichat/event/ToDo.java
@@ -1,4 +1,4 @@
-package hiChat.event;
+package hichat.event;
 
 public  class ToDo extends Task{
     public ToDo (String task){
